More reliable code. State saved and read now
Code now uses path of the python code. In principal, now it can be saved
in any place (before it could only be saved in Documents).
File is rewritten when recompiled

The state of sampling is saved in a hidden file. Every time the buttons
are changed, the files is rewritten.